Updated for API v1.3
Small change to the way the Ambience API handles the set
NotificationLaunchActivity method in the Ambience Class. The method now
takes in an Intent-Filter action string instead of a package name and
activity name.

The AmbientService was also update to accommodate the changes to the
NotificationLaunchActivity method. Notifications created by the
AmbientService now append a bundle object containing the current
playing track, playlist, shuffle mode, repeat mode, volume level and
Intent-Filter action string. This information can be used by the user
app when the intent launches the intended activity.